Improve TMT_MSG_SETMODE
Previously, there was no good way to tell which arguments were really
set when doing a mode set.  We now set all the unused ones to -1, which
is probably not valid.  The application can look through 8 arguments
and see the unset ones.

It's still a bit of a hack, but it's better.