Add scrolling region and VT100-like line end behavior
This implements CSI<L>;<H>r, which lets a program specify a range
of lines to scroll, leaving others alone.

It also implements code *similar* to how VT terminals deal with it
when a character is written to the last column in a line.  It
essentially defers actually advancing the line, because there
are cases where we don't want to (at least not yet), e.g., when
the cursor wraps onto the next line and immediately does a CR.
The behavior is not 100% the same as a VT, but it is close, and
I haven't yet noticed bad behavior.